add shared test harness for LS
